Creacion y lista de ordenes

+++ app/domain/repository/repository.go
@@ -15,9 +15,9 @@ type Order interface {
 type Client interface {
 	Save(*model.Client) error
 	FindByName(string) (*model.Client, error)
-	GetById(id int64) model.Client
+	GetById(id uint) model.Client
 	//TODO: Convertir esta lista a una lista de apuntadores "[]*model.Client"
 	All() ([]model.Client, error)
-	Detele(id int64) error
+	Detele(id uint) error
 	Update(*model.Client) error
 }
